PerfTest should have had a float
git-svn-id: http://protobuf-net.googlecode.com/svn/trunk@10 b72047a4-3652-0410-9617-b3994939e97b

+++ Examples/SimpleStream/SimpleStreamDemo.cs
@@ -65,7 +65,7 @@ namespace Examples.SimpleStream
         public string Bar { get; set; }
 
         [DataMember(Order = 3)]
-        public double Blip{ get; set; }
+        public float Blip{ get; set; }
 
         [DataMember(Order = 4)]
         public double Blop { get; set; }
